adds conditional rendering for whether the user has logged in
