+++ src/components/pages/MyAccount/MyAccount.module.css
@@ -1,5 +1,6 @@
 .container {
     display: flex;
+    flex-direction: column;
     justify-content: center;
 align-items: center;
 text-align: center;
