+++ src/components/pages/MyArticles/MyArticles.module.css
@@ -1,8 +1,8 @@
 .container {
     display: flex;
+    flex-direction: column;
     justify-content: center;
 align-items: center;
 text-align: center;
-height: calc(100vh - 60px);
-min-height: 520px;
+min-height: calc(100vh - 60px);
 }
